# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "voting-api",
        "description": "This is a voting-api development server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "peter.hrobar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/questions": {
            "get": {
                "tags": [
                    "no-auth",
                    "Question"
                ],
                "summary": "Show all questions",
                "description": "Show all questions registered in the database",
                "operationId": "showAllQuestions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OAuth",
                    "Question"
                ],
                "summary": "Create a question",
                "description": "Creates a new question.",
                "operationId": "createQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question_text": {
                                        "type": "string",
                                        "example": "What is your favorite color?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Question successfully created"
                                        },
                                        "question": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The question_text field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/questions/{question_id}": {
            "delete": {
                "tags": [
                    "OAuth",
                    "Question"
                ],
                "summary": "Delete a question",
                "description": "Deletes a specific question.",
                "operationId": "deleteQuestion",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Question deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Question not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Question": {
                "title": "Question model",
                "description": "Question model",
                "required": [
                    "question_text"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Auto-incrementing ID of the question",
                        "type": "integer",
                        "format": "integer"
                    },
                    "question_text": {
                        "title": "Question text",
                        "description": "Question text",
                        "type": "string",
                        "format": "string"
                    },
                    "is_closed": {
                        "title": "Is closed?",
                        "description": "Determines if the Question is closed for modification or voting",
                        "type": "boolean",
                        "default": false
                    },
                    "number_of_votes": {
                        "title": "Number of votes",
                        "description": "Number of voting options the question has (calculated field)",
                        "type": "integer",
                        "minimum": 0
                    },
                    "last_vote_at": {
                        "title": "Last date and time when a vote was updated",
                        "description": "Date and time when the last vote was received for the question (calculated field)",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-06-18 12:01:01"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Date and time when the question was created",
                        "type": "string",
                        "format": "datetime",
                        "default": "now",
                        "example": "2023-06-18 12:01:01"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Date and time when the question was updated",
                        "type": "string",
                        "format": "datetime",
                        "default": "now",
                        "example": "2023-06-18 12:01:01"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Question"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}